requirements.txt
# Azure SDKs for authentication and Key Vault
azure-identity>=1.15.0
azure-keyvault-secrets>=4.7.0

# Azure SDK for Container Apps (ACA launcher)
azure-mgmt-appcontainers>=2.0.0

# Azure SDK for Resource Management (helper functions)
azure-mgmt-resource>=23.0.0
